requirements.txt
requests
bs4
google_api_translate
